+++ clx_task_management/models/request_form.py
@@ -3,7 +3,9 @@
 # See LICENSE file for full copyright & licensing details.
 from odoo import fields, api, models, _
 from odoo.exceptions import UserError
+from dateutil.relativedelta import relativedelta
 import datetime
+from pytz import timezone
 
 
 class RequestForm(models.Model):
@@ -14,7 +16,7 @@ class RequestForm(models.Model):
 
     name = fields.Char(string='Name', copy=False)
     partner_id = fields.Many2one('res.partner', string='Customer')
-    request_date = fields.Datetime('Request Submission Date', default=datetime.datetime.today())
+    request_date = fields.Datetime('Request Submission Date', default=datetime.datetime.today(), copy=False)
     description = fields.Text('Project Title',
                               help="Choose a title for this client’s project request")
     request_line = fields.One2many('request.form.line', 'request_form_id',
@@ -136,30 +138,11 @@ class RequestForm(models.Model):
         """
         Prepared vals for sub task
         :param sub_task: browsable object of the sub.task
-        :param main_task: browsable object of the main.task
+        :param main_task: browsable object of the project.task
         :param line: browsable object of the request.line
         :return: dictionary for the sub task
         """
         stage_id = self.env.ref('clx_task_management.clx_project_stage_1')
-        today = fields.Date.today()
-        if line.request_form_id.priority == 'high':
-            if line.req_type in ('update', 'budget'):
-                business_days_to_add = 1
-            else:
-                business_days_to_add = 3
-        else:
-            if line.req_type in ('update', 'budget'):
-                business_days_to_add = 3
-            else:
-                business_days_to_add = 5
-        current_date = today
-        # code for skip saturday and sunday for set deadline on task.
-        while business_days_to_add > 0:
-            current_date += datetime.timedelta(days=1)
-            weekday = current_date.weekday()
-            if weekday >= 5:  # sunday = 6, saturday = 5
-                continue
-            business_days_to_add -= 1
         if stage_id:
             vals = {
                 'name': sub_task.sub_task_name,
@@ -170,7 +153,7 @@ class RequestForm(models.Model):
                 'sub_task_id': sub_task.id,
                 'team_ids': sub_task.team_ids.ids,
                 'team_members_ids': sub_task.team_members_ids.ids,
-                'date_deadline': self.intended_launch_date if self.intended_launch_date else current_date,
+                'date_deadline': main_task.date_deadline,
                 'tag_ids': sub_task.tag_ids.ids if sub_task.tag_ids else False,
                 'account_user_id': main_task.project_id.partner_id.account_user_id.id if main_task.project_id.partner_id.account_user_id else False,
                 'clx_priority': main_task.project_id.priority,
@@ -189,16 +172,30 @@ class RequestForm(models.Model):
         """
         stage_id = self.env.ref('clx_task_management.clx_project_stage_1')
         today = fields.Date.today()
+        current_day_with_time = self.write_date
+        user_tz = self.env.user.tz or 'US/Pacific'
+        current_day_with_time = timezone('UTC').localize(current_day_with_time).astimezone(timezone(user_tz))
+        date_time_str = today.strftime("%d/%m/%y")
+        date_time_str += ' 14:00:00'
+        comparsion_date = datetime.datetime.strptime(date_time_str, '%d/%m/%y %H:%M:%S')
+        # if current_day_with_time after 2 pm:
+        #     today + 1 day
+        if current_day_with_time.time() > comparsion_date.time():
+            business_days_to_add = 1
+        else:
+            business_days_to_add = 0
+
         if line.request_form_id.priority == 'high':
             if line.req_type in ('update', 'budget'):
-                business_days_to_add = 1
+                business_days_to_add += 0
             else:
-                business_days_to_add = 3
+                business_days_to_add += 2
         else:
             if line.req_type in ('update', 'budget'):
-                business_days_to_add = 3
+                business_days_to_add += 2
             else:
-                business_days_to_add = 5
+                business_days_to_add += 4
+
         current_date = today
         # code for skip saturday and sunday for set deadline on task.
         while business_days_to_add > 0:
@@ -233,26 +230,15 @@ class RequestForm(models.Model):
         :Param : partner_id : browsable object of the partner
         :return : return dictionary
         """
-        today = fields.Date.today()
-        if all(line.req_type == 'update' for line in self.request_line):
-            business_days_to_add = 3
-        else:
-            business_days_to_add = 5
-        current_date = today
-        # code for skip saturday and sunday for set deadline on task.
-        while business_days_to_add > 0:
-            current_date += datetime.timedelta(days=1)
-            weekday = current_date.weekday()
-            if weekday >= 5:  # sunday = 6, saturday = 5
-                continue
-            business_days_to_add -= 1
+        date_list = self.calculated_date(self.request_line)
+        max_date = max(date_list)
         vals = {
             'partner_id': partner_id.id,
             'name': description,
             'clx_state': 'new',
             'clx_sale_order_ids': self.sale_order_id.ids if self.sale_order_id.ids else False,
             'user_id': self.partner_id.account_user_id.id if self.partner_id.account_user_id else False,
-            'deadline': self.intended_launch_date if self.intended_launch_date else current_date,
+            'deadline': max_date,
             'priority': self.priority,
             'clx_attachment_ids': self.clx_attachment_ids.ids
         }
@@ -272,18 +258,39 @@ class RequestForm(models.Model):
 
     def calculated_date(self, line):
         date_list = []
-        for l in line:
+        for line in line:
             today = fields.Date.today()
-            if l.request_form_id.priority == 'high':
-                if l.req_type in ('update', 'budget'):
-                    business_days_to_add = 1
+            current_day_with_time = self.write_date
+            user_tz = self.env.user.tz or 'US/Pacific'
+            current_day_with_time = timezone('UTC').localize(current_day_with_time).astimezone(timezone(user_tz))
+            date_time_str = today.strftime("%d/%m/%y")
+            date_time_str += ' 14:00:00'
+            comparsion_date = datetime.datetime.strptime(date_time_str, '%d/%m/%y %H:%M:%S')
+            # if current_day_with_time after 2 pm:
+            #     today + 1 day
+            if current_day_with_time.time() > comparsion_date.time():
+                today = today + relativedelta(days=1)
+                if line.request_form_id.priority == 'high':
+                    if line.req_type in ('update', 'budget'):
+                        business_days_to_add = 1
+                    else:
+                        business_days_to_add = 3
                 else:
-                    business_days_to_add = 3
+                    if line.req_type in ('update', 'budget'):
+                        business_days_to_add = 3
+                    else:
+                        business_days_to_add = 5
             else:
-                if l.req_type in ('update', 'budget'):
-                    business_days_to_add = 3
+                if line.request_form_id.priority == 'high':
+                    if line.req_type in ('update', 'budget'):
+                        business_days_to_add = 0
+                    else:
+                        business_days_to_add = 2
                 else:
-                    business_days_to_add = 5
+                    if line.req_type in ('update', 'budget'):
+                        business_days_to_add = 2
+                    else:
+                        business_days_to_add = 4
             current_date = today
             # code for skip saturday and sunday for set deadline on task.
             while business_days_to_add > 0:
@@ -303,7 +310,6 @@ class RequestForm(models.Model):
         self.ensure_one()
         if self.request_line and self.intended_launch_date:
             date_list = self.calculated_date(self.request_line)
-            print(date_list)
             max_date = max(date_list)
             if self.intended_launch_date < max_date:
                 raise UserError('Please check Intended launch Date !!')
@@ -316,7 +322,7 @@ class RequestForm(models.Model):
             raise UserError('Please add Project Title!!')
         if not self.request_line:
             raise UserError('There is no Request Line, Please add some line')
-        if self.request_line and any(not line.description for line in self.request_line):
+        if self.request_line and any(not line.description or not line.task_id for line in self.request_line):
             raise UserError(
                 'Please add some Instruction on every line If do not have description Please delete that line.')
 
@@ -437,7 +443,9 @@ class RequestFormLine(models.Model):
                 raise UserError(_(
                     """There is no subscription available for this customer."""))
         if not self.category_id:
-            return {'domain': {'task_id': [('req_type', '=', self.req_type)]}}
+            return {'domain': {'task_id': [('req_type', '=', self.req_type),
+                                           ('pull_to_request_form', '=', True)]}}
         elif self.category_id:
             return {'domain': {'task_id': [('req_type', '=', self.req_type),
-                                           ('category_id', '=', self.category_id.id)]}}
+                                           ('category_id', '=', self.category_id.id),
+                                           ('pull_to_request_form', '=', True)]}}