+++ RELEASE_NOTES.md
@@ -1364,3 +1364,88 @@
 ## 02/15/2021
 #### Version 13.1.1.0.37
 - [FIX] contact_modification : added access rights for the contact creation.
+
+## 02/16/2021
+#### Version 13.1.0.0.68
+- [FIX] clx_invoice_policy : Fix the yearly subscription creation from the generate invoice with date range.
+
+## 02/17/2021
+#### Version 13.3.0.0.97
+- [FIX] clx_task_management : added launch date calculation on task.
+
+## 02/18/2021
+#### Version 13.1.0.0.69
+- [FIX] clx_invoice_policy : Fixed the issue of generating random invoices.
+
+
+## 02/19/2021
+#### Version 13.1.0.0.70
+- [FIX] clx_invoice_policy : Fixed the label generating issue.
+
+## 02/19/2021
+#### Version 13.0.1.0.1
+- [ADD] mail_optional_follower_notification : Initial commit for the not send automail to followers. 
+
+## 02/22/2021
+#### Version 13.0.1.0.2
+- [FIX] mail_optional_follower_notification : Fixed the issue related to config setting issue.
+
+## 02/22/2021
+#### Version 13.1.0.0.71
+- [FIX] clx_invoice_policy : Fixed the issue subscription lines in not updated on invoice.
+
+## 02/22/2021
+#### Version 13.3.0.0.98
+- [FIX] clx_task_management : fixed the timezone issue in deadline of project and task.
+
+## 02/22/2021
+#### Version 13.1.0.0.72
+- [FIX] clx_invoice_policy : Added prorate end amount for invoice creation.
+
+## 02/22/2021
+#### Version 13.1.0.0.73
+- [FIX] clx_invoice_policy : fix the date comparsion issue.
+
+## 02/23/2021
+#### Version 13.1.0.0.74
+- [FIX] clx_invoice_policy : fixed the end date issue while creation manualy invoice.
+
+## 02/22/2021
+#### Version 13.3.0.0.99
+- [FIX] clx_task_management : Fixed the deadline related issue on subtask.
+
+## 02/23/2021
+#### Version 13.1.0.0.75
+- [FIX] clx_invoice_policy : Fixed invoice generation issue when user select more than 1 one month data range.
+
+## 02/23/2021
+#### Version 13.3.0.0.100
+- [FIX] clx_task_management : Fixed deadline issue when user create manually parent task.
+
+## 02/24/2021
+#### Version 13.3.0.1.1
+- [FIX] clx_task_management : fixed project archieved issue.
+
+## 02/24/2021
+#### Version 13.1.0.0.76
+- [FIX] clx_invoice_policy : added description on invoice
+
+## 02/24/2021
+#### Version 13.3.0.0.32
+- [FIX] clx_invoice_reports : added description on invoice pdf reports.
+
+## 02/24/2021
+#### Version 13.0.1.0.3
+- [FIX] mail_optional_follower_notification : changed in sending mail flow.
+
+## 02/25/2021
+#### Version 13.3.0.1.2
+- [FIX] clx_task_management : Fixed the deadline issue, the deadline was not set on subtask while creating manually task.
+
+## 02/25/2021
+#### Version 13.1.0.0.77
+- [FIX] clx_invoice_policy : Fixed the issue of duplicate generating issue.
+
+## 02/25/2021
+#### Version 13.1.0.0.78
+- [ADD] clx_invoice_policy : Merger 2 sale order lines in invoice if those lines have same description.